# Conflicts:
#	app/src/main/java/com/autohub/skln/models/tutor/TutorAcademicInfo.kt
#	app/src/main/java/com/autohub/skln/models/tutor/TutorData.kt
#	app/src/main/java/com/autohub/skln/models/tutor/TutorLocation.kt
#	app/src/main/java/com/autohub/skln/models/tutor/TutorOtpPassword.kt
#	app/src/main/java/com/autohub/skln/models/tutor/TutorPackageInfo.kt
#	app/src/main/java/com/autohub/skln/models/tutor/TutorPersonalInfo.kt
#	app/src/main/java/com/autohub/skln/models/tutor/TutorQualification.kt
#	app/src/main/java/com/autohub/skln/models/tutormodels/TutorAcademicInfo.kt
#	app/src/main/java/com/autohub/skln/models/tutormodels/TutorData.kt
#	app/src/main/java/com/autohub/skln/models/tutormodels/TutorLocation.kt
#	app/src/main/java/com/autohub/skln/models/tutormodels/TutorOtpPassword.kt
#	app/src/main/java/com/autohub/skln/models/tutormodels/TutorPackageInfo.kt
#	app/src/main/java/com/autohub/skln/models/tutormodels/TutorPersonalInfo.kt
#	app/src/main/java/com/autohub/skln/models/tutormodels/TutorQualification.kt
#	tutormodule/src/main/java/com/autohub/tutormodule/data/modal/TutorAcademicInfo.kt
#	tutormodule/src/main/java/com/autohub/tutormodule/data/modal/TutorData.kt
#	tutormodule/src/main/java/com/autohub/tutormodule/data/modal/TutorLocation.kt
#	tutormodule/src/main/java/com/autohub/tutormodule/data/modal/TutorOtpPassword.kt
#	tutormodule/src/main/java/com/autohub/tutormodule/data/modal/TutorPackageInfo.kt
#	tutormodule/src/main/java/com/autohub/tutormodule/data/modal/TutorPersonalInfo.kt
#	tutormodule/src/main/java/com/autohub/tutormodule/data/modal/TutorQualification.kt
#	tutormodule/src/main/java/com/autohub/tutormodule/ui/dashboard/DashboardActivity.kt

+++ studentmodule/src/main/java/com/autohub/studentmodule/models/ExploreFilter.kt
@@ -4,4 +4,4 @@ package com.autohub.studentmodule.models
  * Created by Vt Netzwelt
  */
 data class ExploreFilter(var studentClass: String = "", var subjectName: String = "",
-                         var latitude: Float = 0.toFloat(), var longitude: Float = 0.toFloat(), var filterType : String)
+                         var latitude: Double = 0.0, var longitude: Double = 0.0, var filterType: String)
